More logging and DryRun for NX tasks
* Add logging for executed Nx commands

Log the Nx command being executed to improve traceability and debugging. This helps in understanding the execution flow and identifying potential issues more easily.

* Refactor CI build tasks and update Playwright config

Set `reuseExistingServer` to `false` in Playwright config for consistency. Replaced `runTargetForAffected` with `runTargetsForAll` in CI tasks to target all projects instead of only affected ones. Added support for `runTargetsForAll` in `nx-tasks` and enabled `--dryRun` for release publishing commands.

* Add support for dry-run mode in nx-tasks

Included a check for the `dryRun` context to append the `--dry-run` flag when applicable. This ensures safer task execution during development or testing without making actual changes.

* Add CI environment detection and logging in build context

Introduced a method to determine if the build is running in a CI environment by checking the GITHUB_ACTIONS environment variable. Added logging in the build executor to output whether the build is running on a CI build server. This improves transparency and environment awareness during builds.